Fixed some minor serializing warnings
git-svn-id: svn+ssh://scheduler.csc.calpoly.edu/repo@1238 13b268f7-8392-4c9d-bc7d-1e80a86bc2f9

+++ implementation/model/java/edu/calpoly/csc/scheduler/model/db/idb/CoursePreference.java
@@ -14,7 +14,7 @@ import edu.calpoly.csc.scheduler.model.db.cdb.*;
  */
 public class CoursePreference implements Serializable
 {
-	public static final int serialVersionUID = 42;
+	public static final long serialVersionUID = 42;
 
 	/** The prefered course */
 	private Course course;
